feat: moved three dots spinner into own atom
