+++ src/components/atoms/Spinner/Spinner.md
@@ -27,11 +27,3 @@ import { Spinner } from '@zopauk/react-components';
 
 <Spinner styling="negative" />;
 ```
-
-ThreeDotsSpinner:
-
-```ts
-import { ThreeDotsSpinner } from '@zopauk/react-components';
-
-<ThreeDotsSpinner />;
-```
