+++ src/components/atoms/ThreeDotsSpinner/ThreeDotsSpinner.md
@@ -1,0 +1,11 @@
+### Summary
+
+Use `<ThreeDotsSpinner />` whenever we need to notify the user that something is loading.
+
+### Example
+
+```ts
+import { ThreeDotsSpinner } from '@zopauk/react-components';
+
+<ThreeDotsSpinner />;
+```
